DAL And TaskService Layer

+++ TaskManager/TaskDomain/DomainModel/LoginUserDm.cs
@@ -1,17 +1,13 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
 
 namespace TaskDomain.DomainModel
 {
     public class LoginUserDm
     {
-        public long Id { get; set; }
-        public long RoleId { get; set; }
-        public long EmpId { get; set; }
+        public int Id { get; set; }
+        public int RoleId { get; set; }
+        public int EmpId { get; set; }
 
         [Required(ErrorMessage = "Please Enter Email Address")]
         [DataType(DataType.EmailAddress)]
